+++ src/client/mission_client.rs
@@ -1,10 +1,5 @@
-use std::future::ready;
-
-use futures_util::{Stream, StreamExt};
 use stubs::mission;
 use stubs::mission::v0::mission_service_client::MissionServiceClient;
-use stubs::mission::v0::stream_events_response::Event;
-use tonic::Status;
 
 use super::{request_with_deadline, GrpcChannel, GrpcResult};
 
@@ -41,29 +36,6 @@ impl MissionClient {
             .map_err(Box::new)?
             .into_inner();
         Ok(res.datetime)
-    }
-
-    pub async fn stream_events(
-        &mut self,
-    ) -> GrpcResult<impl Stream<Item = Result<(f64, Event), Status>>> {
-        let events = self
-            .svc
-            .stream_events(request_with_deadline(mission::v0::StreamEventsRequest {}))
-            .await
-            .map_err(Box::new)?
-            .into_inner()
-            .filter_map(|event| {
-                ready(match event {
-                    Ok(stubs::mission::v0::StreamEventsResponse {
-                        time,
-                        event: Some(event),
-                        ..
-                    }) => Some(Ok((time, event))),
-                    Err(err) => Some(Err(err)),
-                    Ok(_) => None,
-                })
-            });
-        Ok(events)
     }
 
     pub async fn get_session_id(&mut self) -> GrpcResult<i64> {